backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
Flask-CORS
pyotp
cryptography
python-dotenv
Werkzeug 